Move #create and #destroy actions to GymMembershipsController
