+++ app/controllers/gyms_controller.rb
@@ -13,28 +13,4 @@ class GymsController < ApplicationController
     @gym_friends = BackEndFacade.get_friends_at_gym(params[:id], current_user.id)
     @gym_non_friends = BackEndFacade.get_non_friends_at_gym(yelp_gym_id: params[:id], user_id: current_user.id)
   end
-
-  def create
-    BackEndFacade.create_gym_membership(gym_membership_params)
-    redirect_to dashboard_path(current_user.id)
-  end
-
-  def destroy
-    BackEndFacade.delete_gym_membership(gym_membership_destroy_params)
-    redirect_to dashboard_path(current_user.id)
-    flash[:alert] = 'Gym removed... but you should probably join a new one ASAP if you want those gains!'
-  end
-
-  private
-
-  def gym_membership_params
-    params.permit(:user_id, :yelp_gym_id, :gym_name)
-  end
-
-  def gym_membership_destroy_params
-    {
-      user_id: params[:user_id],
-      id: params[:id]
-    }
-  end
 end
